DType Int64 -> Go int64 convertion type fixes.

+++ examples/cifar/cifar.go
@@ -130,7 +130,7 @@ func LoadCifar10(manager *Manager, baseDir string, dtype shapes.DType) (partitio
 	labels := tensor.FromShape(shapes.Make(shapes.I64, NumExamples, 1))
 	labelsRef := labels.AcquireData()
 	defer labelsRef.Release()
-	labelsData := labelsRef.Flat().([]int)
+	labelsData := labelsRef.Flat().([]int64)
 
 	var labelImageBytes [imageSizeBytes + 1]byte
 	for fileIdx := 0; fileIdx < 6; fileIdx++ {
@@ -165,7 +165,7 @@ func LoadCifar10(manager *Manager, baseDir string, dtype shapes.DType) (partitio
 			if err != nil {
 				panic(errors.WithMessagef(err, "failed converting bytes to tensor of %s", dtype))
 			}
-			labelsData[exampleIdx] = int(labelImageBytes[0])
+			labelsData[exampleIdx] = int64(labelImageBytes[0])
 		}
 	}
 	return partitionImagesAndLabels(manager, images, labels)
